Remove not needed includes
Also:
    Change esp32 partition size
    Code clean up
    Bluetooth and WiFi services in ConnectionHandler

+++ src/connection/drv/Bluetooth/DeviceCallbackHandler.cpp
@@ -45,7 +45,7 @@ Command translateStringToCommand(String commandString)
 
 DeviceCallbackHandler::DeviceCallbackHandler(char *accessPointName,
                                              std::shared_ptr<BLECharacteristic> &characteristic,
-                                             std::uint32_t jsonBufferSize)
+                                             std::uint16_t jsonBufferSize)
         : accessPointName(accessPointName)
         , jsonBuffer(jsonBufferSize)
         , characteristics(characteristic)
